Updated the linear charging time
Charging time updated to be more realistic
TODO need to change the order when other agents are closer to a charging station

+++ Edge.java
@@ -3,6 +3,6 @@ import java.io.Serializable;
 public class Edge{
     protected Vertex source;
     protected Vertex destination;
-    protected int cost;
+    protected double cost;
 
 }
